refactor(substrate): move the id-level numeric value tower into sparq-substrate (sq-ev41x) [OPUS-4.8]
Phase 2 of the shared-eval-substrate move-chain (epic sq-qonbz,
research/shared-eval-substrate.md, Option C). Behaviour-neutral code-move: the
engine's id-level XSD numeric value tower moves OUT of sparq-engine::exec INTO
the sparq-substrate leaf crate, and sparq-engine now CONSUMES it. Zero semantic
change.

Moved verbatim into `sparq_substrate::numeric` (feature `numeric`, default-off):
- `Num` (Int/Dec/Float/Double) + all value methods: rank, f64, to_dec, binop,
  neg, abs, ceil, floor, round, datatype, lexical, canonical_lexical, is_nan,
  is_zero, of_literal
- `as_numeric(&Literal)` — the free-function classifier (= Num::of_literal)
- `Dec` (exact fixed-point) + parse/parse_lexical/align/checked_{add,sub,mul,div}/
  cmp/round_to_int/f64/lexical
- `ArithOp`, `RoundMode`, and the XSD lexical helpers split_decimal,
  parse_xsd_f64, parse_xsd_f32, apply_f64, fmt_xsd_double

CRITICAL RECONCILIATION: the #1290 scaffold's `as_numeric` was a PLACEHOLDER
(Dec::parse + bare v.parse::<f32/f64>()). This move replaces it with the engine's
EXACT lexical path — Dec::parse_lexical (scale-preserving) for xsd:decimal and
parse_xsd_f32/f64 (XSD INF/-INF/NaN/exponent spellings) for float/double — so
classification is bit-identical to pre-move.

Engine-side: `as_numeric(&Value)` and `num_canonical_term(Num)->Value` stay in
exec.rs (they touch the engine-private `Value`); everything else is imported from
the substrate under the same names, so every call site is unchanged. The
`compare_values` total order is NOT moved here — it is irreducibly coupled to the
engine's `Value` enum + the temporal subsystem, so it moves with `Value` in a
follow-up (scoped honestly; see PR body).

ZERO-OVERHEAD: no Box<dyn>/&dyn/vtable in the moved module (grep + clippy clean).
Every kernel carries #[inline]; cross-crate inlining holds under the workspace
fat-LTO release profile (measured: plain #[inline] beats both #[inline(always)]
and no-inline on the wasm bundle).

Gates (both feature states): cargo build + clippy --all-targets -D warnings +
cargo +1.88 check GREEN; substrate `numeric` tests GREEN; full W3C SPARQL
conformance 1225 pass / 0 fail / 4 documented-divergence — pass+divergence 1229,
== the committed ratchet floor, bit-identical to pre-move (strict run exits 0);
rustdoc --workspace --all-features -D warnings clean; readme-template 0 deviations;
perf-gate PASS (store_bytes_per_triple=92, dict_bytes_per_term=53 identical —
sparq-core untouched; wasm_bundle_bytes within the 2% band).
