+++ crates/sparq-engine/Cargo.toml
@@ -204,6 +204,14 @@ yannakakis = ["semijoin-bitmap"]
 # `version` alongside `path`: cargo uses the path locally and the version on crates.io
 # (required to publish — crates.io strips `path`).
 sparq-core = { path = "../sparq-core", version = "0.1.0", default-features = false }
+# [OPUS-4.8] sq-ev41x (epic sq-qonbz): the shared id-level numeric value tower — `Num` /
+# `Dec` / `as_numeric` and the XSD lexical helpers — moved out of `exec.rs` into the
+# `sparq-substrate` leaf crate so the engine AND the reasoners can share one definition with
+# NO `Box<dyn>` on the hot path (research/shared-eval-substrate.md, Option C, Phase 2). The
+# engine uses these unconditionally (FILTER / BIND / ORDER BY arithmetic), so the `numeric`
+# feature is enabled here, not optionally — `oxrdf` (which `numeric` classifies) is already
+# in the engine's dep tree, so this pulls no new crate and the lean wasm bundle is unchanged.
+sparq-substrate = { path = "../sparq-substrate", version = "0.1.0", default-features = false, features = ["numeric"] }
 oxrdf.workspace = true
 spargebra.workspace = true
 # Relative-IRI resolution for IRI()/URI() against the query's BASE (already in the
