+++ crates/sparq-engine/src/exec.rs
@@ -13,6 +13,13 @@ use sparq_core::dict::{self, Id, NO_ID};
 use sparq_core::store::Pattern as IdPattern;
 use sparq_core::temporal::{Temporal, Timeline};
 use sparq_core::Graph;
+// [OPUS-4.8] sq-ev41x (epic sq-qonbz): the id-level numeric value tower (`Num` / `Dec` /
+// `ArithOp` / `RoundMode` and the XSD lexical helpers) now lives in the `sparq-substrate`
+// leaf crate so the engine AND the reasoners share ONE definition with no `Box<dyn>` on the
+// hot path (research/shared-eval-substrate.md, Option C, Phase 2). This is a behaviour-neutral
+// code-move: the engine re-imports the same items here under the same names, so every call
+// site below is unchanged and the W3C conformance / ORDER BY / numeric tests are bit-identical.
+use sparq_substrate::numeric::{parse_xsd_f32, parse_xsd_f64, split_decimal, ArithOp, Dec, Num};
 use rustc_hash::FxHashSet;
 use spargebra::algebra::{
     AggregateExpression, AggregateFunction, Expression, GraphPattern, OrderExpression, PropertyPathExpression,
@@ -6187,7 +6194,7 @@ fn eval_aggregate(graph: &Graph, local: &LocalVocab, b: &Bindings, members: &[us
                 // SUM/AVG with operand-type promotion: int+int stays integer, decimals
                 // stay decimal (exact), floats/doubles promote. Any non-numeric or
                 // errored member makes the whole aggregate a type error (-> unbound).
-                AggregateFunction::Sum => Ok(sum_values(&vals, errored).map(Num::canonical_term).unwrap_or(Value::Error)),
+                AggregateFunction::Sum => Ok(sum_values(&vals, errored).map(num_canonical_term).unwrap_or(Value::Error)),
                 AggregateFunction::Avg => {
                     if vals.is_empty() && !errored {
                         return Ok(Value::Num(Num::Int(0))); // AVG({}) = 0 per SPARQL
@@ -6285,7 +6292,7 @@ fn minmax_values(vals: Vec<Value>, keep: Ordering) -> Value {
                     best = n;
                 }
             }
-            best.canonical_term()
+            num_canonical_term(best)
         }
         None => {
             let cmp = |a: &Value, c: &Value| compare_values(a, c).unwrap_or(Ordering::Equal);
@@ -6644,334 +6651,18 @@ enum Value {
     Error,
 }
 
-/// A COMPUTED numeric value carrying its XSD type, implementing the SPARQL/XPath
-/// operand-type-promotion tower: integer < decimal < float < double. Arithmetic
-/// promotes both operands to the greater type; integer and decimal arithmetic is
-/// EXACT (i64 / fixed-point [`Dec`]), falling back to double only on overflow.
-/// Serialisation (see [`Num::lexical`]) is the XSD canonical form of the type.
-#[derive(Clone, Copy, Debug)]
-enum Num {
-    Int(i64),
-    Dec(Dec),
-    Float(f32),
-    Double(f64),
-}
+// [OPUS-4.8] sq-ev41x: `Num` / `Dec` / `ArithOp` / `RoundMode` and the XSD lexical helpers
+// (`apply_f64`, `parse_xsd_f64`, `parse_xsd_f32`, `fmt_xsd_double`) MOVED to
+// `sparq_substrate::numeric` (imported at the top of this file). The only `Num` operation
+// that produced an engine-private `Value` — `canonical_term` — stays here as a free helper,
+// since the substrate must not know about the engine's `Value`/`Term`. Everything else is
+// the substrate's `Num` verbatim, so behaviour is unchanged.
 
-#[derive(Clone, Copy, PartialEq)]
-enum ArithOp {
-    Add,
-    Sub,
-    Mul,
-    Div,
-}
-
-impl Num {
-    /// Promotion rank in the XPath numeric tower.
-    fn rank(self) -> u8 {
-        match self {
-            Num::Int(_) => 0,
-            Num::Dec(_) => 1,
-            Num::Float(_) => 2,
-            Num::Double(_) => 3,
-        }
-    }
-
-    fn f64(self) -> f64 {
-        match self {
-            Num::Int(i) => i as f64,
-            Num::Dec(d) => d.f64(),
-            Num::Float(f) => f as f64,
-            Num::Double(d) => d,
-        }
-    }
-
-    fn to_dec(self) -> Option<Dec> {
-        match self {
-            Num::Int(i) => Some(Dec { mant: i as i128, scale: 0 }),
-            Num::Dec(d) => Some(d),
-            _ => None,
-        }
-    }
-
-    /// The typed numeric value of a literal, or `None` if the literal is not a
-    /// well-formed numeric (an ill-formed numeric operand is a SPARQL type error).
-    fn of_literal(l: &Literal) -> Option<Num> {
-        if l.language().is_some() {
-            return None;
-        }
-        let dt = l.datatype();
-        let v = l.value().trim();
-        if sparq_core::is_integer_datatype(dt.as_str()) {
-            if let Ok(i) = v.parse::<i64>() {
-                return Some(Num::Int(i));
-            }
-            // Integer beyond i64: exact i128 mantissa if it fits (scale 0 = integer
-            // lexical), else not representable -> double.
-            return match Dec::parse(v) {
-                Some(d) if d.scale == 0 => Some(Num::Dec(d)),
-                Some(_) => None, // "1.5"^^xsd:integer is ill-formed
-                None => None,
-            };
-        }
-        if dt == xsd::DECIMAL {
-            return Dec::parse_lexical(v).map(Num::Dec);
-        }
-        if dt == xsd::FLOAT {
-            return parse_xsd_f32(v).map(Num::Float);
-        }
-        if dt == xsd::DOUBLE {
-            return parse_xsd_f64(v).map(Num::Double);
-        }
-        None
-    }
-
-    /// `op` under XPath operand promotion. `None` is a SPARQL type error (exact-type
-    /// division by zero); exact-arithmetic overflow falls back to double, mirroring
-    /// the engine's previous f64 behaviour.
-    fn binop(self, o: Num, op: ArithOp) -> Option<Num> {
-        let rank = self.rank().max(o.rank());
-        if rank == 3 {
-            return Some(Num::Double(apply_f64(self.f64(), o.f64(), op)));
-        }
-        if rank == 2 {
-            let (a, b) = (self.f64() as f32, o.f64() as f32);
-            return Some(Num::Float(apply_f64(a as f64, b as f64, op) as f32));
-        }
-        // Exact tier: integer / decimal.
-        let (a, b) = (self.to_dec()?, o.to_dec()?);
-        if op == ArithOp::Div {
-            // xsd:integer / xsd:integer is DECIMAL division per SPARQL; exact-type
-            // division by zero is a type error (not INF/NaN).
-            if b.mant == 0 {
-                return None;
-            }
-            return match a.checked_div(b) {
-                Some(d) => Some(Num::Dec(d)),
-                None => Some(Num::Double(self.f64() / o.f64())),
-            };
-        }
-        if rank == 0 {
-            // integer op integer -> integer (i64; on overflow fall back to double).
-            let (x, y) = (match self { Num::Int(i) => i, _ => unreachable!() }, match o { Num::Int(i) => i, _ => unreachable!() });
-            let r = match op {
-                ArithOp::Add => x.checked_add(y),
-                ArithOp::Sub => x.checked_sub(y),
-                ArithOp::Mul => x.checked_mul(y),
-                ArithOp::Div => unreachable!(),
-            };
-            return Some(match r {
-                Some(i) => Num::Int(i),
-                None => Num::Double(apply_f64(x as f64, y as f64, op)),
-            });
-        }
-        let r = match op {
-            ArithOp::Add => a.checked_add(b),
-            ArithOp::Sub => a.checked_sub(b),
-            ArithOp::Mul => a.checked_mul(b),
-            ArithOp::Div => unreachable!(),
-        };
-        Some(match r {
-            Some(d) => Num::Dec(d),
-            None => Num::Double(apply_f64(self.f64(), o.f64(), op)),
-        })
-    }
-
-    fn neg(self) -> Num {
-        match self {
-            Num::Int(i) => i.checked_neg().map(Num::Int).unwrap_or(Num::Double(-(i as f64))),
-            Num::Dec(d) => d.mant.checked_neg().map(|m| Num::Dec(Dec { mant: m, scale: d.scale })).unwrap_or(Num::Double(-d.f64())),
-            Num::Float(f) => Num::Float(-f),
-            Num::Double(d) => Num::Double(-d),
-        }
-    }
-
-    fn abs(self) -> Num {
-        match self {
-            Num::Int(i) => i.checked_abs().map(Num::Int).unwrap_or(Num::Double((i as f64).abs())),
-            Num::Dec(d) => d.mant.checked_abs().map(|m| Num::Dec(Dec { mant: m, scale: d.scale })).unwrap_or(Num::Double(d.f64().abs())),
-            Num::Float(f) => Num::Float(f.abs()),
-            Num::Double(d) => Num::Double(d.abs()),
-        }
-    }
-
-    fn ceil(self) -> Num {
-        match self {
-            Num::Int(_) => self,
-            Num::Dec(d) => Num::Dec(d.round_to_int(RoundMode::Ceil)),
-            Num::Float(f) => Num::Float(f.ceil()),
-            Num::Double(d) => Num::Double(d.ceil()),
-        }
-    }
-
-    fn floor(self) -> Num {
-        match self {
-            Num::Int(_) => self,
-            Num::Dec(d) => Num::Dec(d.round_to_int(RoundMode::Floor)),
-            Num::Float(f) => Num::Float(f.floor()),
-            Num::Double(d) => Num::Double(d.floor()),
-        }
-    }
-
-    /// XPath fn:round — round half towards POSITIVE INFINITY (so round(-2.5) = -2),
-    /// preserving the argument's datatype.
-    fn round(self) -> Num {
-        match self {
-            Num::Int(_) => self,
-            Num::Dec(d) => Num::Dec(d.round_to_int(RoundMode::HalfUp)),
-            Num::Float(f) => Num::Float((f + 0.5).floor()),
-            Num::Double(d) => Num::Double((d + 0.5).floor()),
-        }
-    }
-
-    fn datatype(self) -> oxrdf::NamedNodeRef<'static> {
-        match self {
-            Num::Int(_) => xsd::INTEGER,
-            Num::Dec(_) => xsd::DECIMAL,
-            Num::Float(_) => xsd::FLOAT,
-            Num::Double(_) => xsd::DOUBLE,
-        }
-    }
-
-    /// XSD CANONICAL lexical form of the value: integers as plain digits; decimals
-    /// preserving the arithmetic scale ("3.0" for 1.0+2, "3" for CEIL(2.5)); float /
-    /// double in mantissa-E-exponent form with a mandatory fractional digit ("3.21E4",
-    /// "2.0E-1") and NaN / INF / -INF spelled per XSD.
-    fn lexical(self) -> String {
-        match self {
-            Num::Int(i) => i.to_string(),
-            Num::Dec(d) => d.lexical(),
-            // f32 must be formatted as f32 (shortest round-trip); via f64 it would
-            // grow spurious digits ("2.0000000298023224E-1" for 0.2f32).
-            Num::Float(f) => {
-                if f.is_nan() {
-                    "NaN".to_string()
-                } else if f == f32::INFINITY {
-                    "INF".to_string()
-                } else if f == f32::NEG_INFINITY {
-                    "-INF".to_string()
-                } else if f.fract() == 0.0 && f.abs() < 1e15 {
-                    format!("{}", f as i64)
-                } else {
-                    let s = format!("{f:E}");
-                    match s.split_once('E') {
-                        Some((m, e)) if !m.contains('.') => format!("{m}.0E{e}"),
-                        _ => s,
-                    }
-                }
-            }
-            Num::Double(d) => fmt_xsd_double(d),
-        }
-    }
-
-    /// STRICT XSD-canonical lexical: float/double ALWAYS in mantissa-E-exponent form
-    /// ("3.21E4", "1.0E2"), never plain. The W3C aggregate expected results use this
-    /// for MIN/MAX/SUM, while arithmetic results use the plain-integral convention of
-    /// [`Num::lexical`] — the suites were generated by different engines.
-    fn canonical_lexical(self) -> String {
-        match self {
-            Num::Int(_) | Num::Dec(_) => self.lexical(),
-            Num::Float(f) => {
-                if f.is_nan() || f.is_infinite() {
-                    self.lexical()
-                } else {
-                    let s = format!("{f:E}");
-                    match s.split_once('E') {
-                        Some((m, e)) if !m.contains('.') => format!("{m}.0E{e}"),
-                        _ => s,
-                    }
-                }
-            }
-            Num::Double(d) => {
-                if d.is_nan() || d.is_infinite() {
-                    self.lexical()
-                } else {
-                    let s = format!("{d:E}");
-                    match s.split_once('E') {
-                        Some((m, e)) if !m.contains('.') => format!("{m}.0E{e}"),
-                        _ => s,
-                    }
-                }
-            }
-        }
-    }
-
-    /// The value as a TERM in strict canonical form (see [`Num::canonical_lexical`]).
-    fn canonical_term(self) -> Value {
-        Value::Term(Term::Literal(Literal::new_typed_literal(self.canonical_lexical(), self.datatype())))
-    }
-
-    fn is_nan(self) -> bool {
-        match self {
-            Num::Float(f) => f.is_nan(),
-            Num::Double(d) => d.is_nan(),
-            _ => false,
-        }
-    }
-
-    fn is_zero(self) -> bool {
-        match self {
-            Num::Int(i) => i == 0,
-            Num::Dec(d) => d.mant == 0,
-            Num::Float(f) => f == 0.0,
-            Num::Double(d) => d == 0.0,
-        }
-    }
-}
-
-fn apply_f64(a: f64, b: f64, op: ArithOp) -> f64 {
-    match op {
-        ArithOp::Add => a + b,
-        ArithOp::Sub => a - b,
-        ArithOp::Mul => a * b,
-        ArithOp::Div => a / b,
-    }
-}
-
-/// Parse an xsd:float/xsd:double lexical: the XSD spellings of the specials, plus the
-/// ordinary scientific notation Rust's parser shares with XSD. `None` = ill-formed.
-fn parse_xsd_f64(v: &str) -> Option<f64> {
-    match v {
-        "NaN" => Some(f64::NAN),
-        "INF" | "+INF" => Some(f64::INFINITY),
-        "-INF" => Some(f64::NEG_INFINITY),
-        // Rust accepts "inf"/"infinity"/"nan" spellings XSD does not; exclude them.
-        _ if v.bytes().all(|c| c.is_ascii_digit() || matches!(c, b'+' | b'-' | b'.' | b'e' | b'E')) => v.parse::<f64>().ok(),
-        _ => None,
-    }
-}
-
-fn parse_xsd_f32(v: &str) -> Option<f32> {
-    parse_xsd_f64(v).map(|d| d as f32)
-}
-
-/// Float/double serialisation: an INTEGRAL value prints as a plain integer ("6",
-/// "1050" — matching the dominant convention across the W3C expected results, which
-/// mix plain and scientific forms); anything else uses the XSD canonical
-/// mantissa-E-exponent form with a mandatory fractional digit ("2.0E-1", "1.5E1").
-fn fmt_xsd_double(v: f64) -> String {
-    if v.is_nan() {
-        return "NaN".to_string();
-    }
-    if v == f64::INFINITY {
-        return "INF".to_string();
-    }
-    if v == f64::NEG_INFINITY {
-        return "-INF".to_string();
-    }
-    if v.fract() == 0.0 && v.abs() < 1e15 {
-        return format!("{}", v as i64);
-    }
-    let s = format!("{v:E}"); // shortest round-trip mantissa, e.g. "2E-1"
-    match s.split_once('E') {
-        Some((m, e)) if !m.contains('.') => format!("{m}.0E{e}"),
-        _ => s,
-    }
-}
-
-enum RoundMode {
-    Ceil,
-    Floor,
-    HalfUp,
+/// The numeric value as a TERM in strict canonical form (see [`Num::canonical_lexical`]).
+/// Engine-side because it constructs the engine's `Value`; the value/lexical logic lives in
+/// the shared substrate.
+fn num_canonical_term(n: Num) -> Value {
+    Value::Term(Term::Literal(Literal::new_typed_literal(n.canonical_lexical(), n.datatype())))
 }
 
 fn effective_boolean(v: &Value) -> bool {
@@ -7349,22 +7040,6 @@ fn nb_is_zero(int: &str, frac: &str) -> bool {
     int.is_empty() && frac.is_empty()
 }
 
-/// Splits a decimal lexical into (negative, integer-digits, fraction-digits), normalised
-/// (no leading zeros on the integer part, no trailing zeros on the fraction). `None` if
-/// the lexical is not digits with at most one `.`.
-fn split_decimal(s: &str) -> Option<(bool, &str, &str)> {
-    let s = s.trim();
-    let (neg, s) = match s.strip_prefix('-') {
-        Some(r) => (true, r),
-        None => (false, s.strip_prefix('+').unwrap_or(s)),
-    };
-    let (int, frac) = s.split_once('.').unwrap_or((s, ""));
-    if (int.is_empty() && frac.is_empty()) || !int.bytes().chain(frac.bytes()).all(|c| c.is_ascii_digit()) {
-        return None;
-    }
-    Some((neg, int.trim_start_matches('0'), frac.trim_end_matches('0')))
-}
-
 /// Significant decimal digits in a numeric lexical — used to decide whether the f64
 /// sargable path is precision-safe (<= 15 digits round-trips through f64 unambiguously).
 fn sig_digits(s: &str) -> usize {
@@ -7377,174 +7052,6 @@ fn sig_digits(s: &str) -> usize {
         frac.trim_start_matches('0').len()
     } else {
         int.len() + frac.len()
-    }
-}
-
-/// An EXACT fixed-point decimal: `mant * 10^-scale`. Used to evaluate `+ - *` on
-/// integer / `xsd:decimal` operands without f64 rounding (`0.1 + 0.2` is exactly `0.3`),
-/// which the f64 arithmetic path gets wrong. Within `i128` range; overflow → `None` →
-/// the caller falls back to f64. Division and `xsd:double`/`float` stay f64.
-#[derive(Clone, Copy, Debug)]
-struct Dec {
-    mant: i128,
-    scale: u32,
-}
-
-impl Dec {
-    /// Parses an integer / decimal lexical (`[+-]?digits(.digits)?`), `None` otherwise.
-    fn parse(s: &str) -> Option<Dec> {
-        let (neg, int, frac) = split_decimal(s)?;
-        let scale = frac.len() as u32;
-        let mut mag: i128 = 0;
-        for &ch in int.as_bytes().iter().chain(frac.as_bytes()) {
-            mag = mag.checked_mul(10)?.checked_add((ch - b'0') as i128)?;
-        }
-        Some(Dec { mant: if neg { -mag } else { mag }, scale })
-    }
-
-    /// Both mantissas scaled to the common (max) scale, or `None` on overflow.
-    fn align(self, o: Dec) -> Option<(i128, i128)> {
-        let scale = self.scale.max(o.scale);
-        let a = self.mant.checked_mul(10i128.checked_pow(scale - self.scale)?)?;
-        let b = o.mant.checked_mul(10i128.checked_pow(scale - o.scale)?)?;
-        Some((a, b))
-    }
-
-    fn checked_add(self, o: Dec) -> Option<Dec> {
-        let (a, b) = self.align(o)?;
-        Some(Dec { mant: a.checked_add(b)?, scale: self.scale.max(o.scale) })
-    }
-    fn checked_sub(self, o: Dec) -> Option<Dec> {
-        let (a, b) = self.align(o)?;
-        Some(Dec { mant: a.checked_sub(b)?, scale: self.scale.max(o.scale) })
-    }
-    fn checked_mul(self, o: Dec) -> Option<Dec> {
-        Some(Dec { mant: self.mant.checked_mul(o.mant)?, scale: self.scale.checked_add(o.scale)? })
-    }
-    fn cmp(self, o: Dec) -> Option<Ordering> {
-        let (a, b) = self.align(o)?;
-        Some(a.cmp(&b))
-    }
-
-    /// Parses an integer / decimal lexical PRESERVING the written scale ("1.0" keeps
-    /// scale 1), unlike [`Dec::parse`] which normalises trailing fraction zeros away.
-    /// The scale is what XSD-canonical serialisation of decimal arithmetic preserves
-    /// (`1.0 + 2` is `"3.0"`), so typed values must carry it.
-    fn parse_lexical(s: &str) -> Option<Dec> {
-        let s = s.trim();
-        let (neg, s) = match s.strip_prefix('-') {
-            Some(r) => (true, r),
-            None => (false, s.strip_prefix('+').unwrap_or(s)),
-        };
-        let (int, frac) = s.split_once('.').unwrap_or((s, ""));
-        if (int.is_empty() && frac.is_empty()) || !int.bytes().chain(frac.bytes()).all(|c| c.is_ascii_digit()) {
-            return None;
-        }
-        let mut mag: i128 = 0;
-        for &ch in int.as_bytes().iter().chain(frac.as_bytes()) {
-            mag = mag.checked_mul(10)?.checked_add((ch - b'0') as i128)?;
-        }
-        Some(Dec { mant: if neg { -mag } else { mag }, scale: frac.len() as u32 })
-    }
-
-    /// EXACT decimal division. The result's scale is the SMALLEST `s >= 1` at which the
-    /// quotient terminates (`0 / 2 = "0.0"`, `11.1 / 5 = "2.22"`); a non-terminating
-    /// quotient is rounded half-up at scale 18. `None` on overflow (caller falls back
-    /// to double); the caller must reject a zero divisor first (type error).
-    fn checked_div(self, o: Dec) -> Option<Dec> {
-        debug_assert!(o.mant != 0);
-        let neg = (self.mant < 0) != (o.mant < 0);
-        let n0 = self.mant.unsigned_abs();
-        let d = o.mant.unsigned_abs();
-        // mant(s) = n0 * 10^(s + o.scale - self.scale) / d
-        let num_den = |s: u32| -> Option<(u128, u128)> {
-            let e = s as i32 + o.scale as i32 - self.scale as i32;
-            if e >= 0 {
-                Some((n0.checked_mul(10u128.checked_pow(e as u32)?)?, d))
-            } else {
-                Some((n0, d.checked_mul(10u128.checked_pow((-e) as u32)?)?))
-            }
-        };
-        const MAX_SCALE: u32 = 18;
-        for s in 1..=MAX_SCALE {
-            let (num, den) = num_den(s)?;
-            if num % den == 0 {
-                let mant = i128::try_from(num / den).ok()?;
-                return Some(Dec { mant: if neg { -mant } else { mant }, scale: s });
-            }
-        }
-        // Non-terminating: round half-up at the max scale.
-        let (num, den) = num_den(MAX_SCALE)?;
-        let q = num / den + u128::from(num % den * 2 >= den);
-        let mant = i128::try_from(q).ok()?;
-        Some(Dec { mant: if neg { -mant } else { mant }, scale: MAX_SCALE })
-    }
-
-    /// Rounds to an integer-valued decimal (scale 0), preserving the decimal TYPE
-    /// (`CEIL("2.5"^^xsd:decimal)` is `"3"^^xsd:decimal`).
-    fn round_to_int(self, mode: RoundMode) -> Dec {
-        if self.scale == 0 || self.mant == 0 {
-            return Dec { mant: self.mant, scale: 0 };
-        }
-        // [OPUS-4.8] `10i128.pow(self.scale)` overflows (debug panic / release wrap) for any
-        // valid decimal whose scale is >= 39 (10^39 > i128::MAX). When the power exceeds i128
-        // the integer part is necessarily 0 (|mant| < i128::MAX < 10^scale), so |value| < 1 and
-        // also < 0.5 (2*i128::MAX < 10^39 <= 10^scale), making the rounded result obvious from
-        // the sign alone — derive it directly instead of constructing the overflowing power.
-        let mant = match 10i128.checked_pow(self.scale) {
-            Some(p) => {
-                let q = self.mant.div_euclid(p);
-                let r = self.mant.rem_euclid(p); // 0..p
-                match mode {
-                    RoundMode::Floor => q,
-                    RoundMode::Ceil => q + i128::from(r > 0),
-                    RoundMode::HalfUp => q + i128::from(r * 2 >= p),
-                }
-            }
-            None => match mode {
-                // |value| < 1: floor of a tiny positive is 0, of a tiny negative is -1.
-                RoundMode::Floor => -i128::from(self.mant < 0),
-                // ceil of a tiny positive is 1, of a tiny negative is 0.
-                RoundMode::Ceil => i128::from(self.mant > 0),
-                // |value| < 0.5 always at this scale, so half-up rounds to 0.
-                RoundMode::HalfUp => 0,
-            },
-        };
-        Dec { mant, scale: 0 }
-    }
-
-    fn f64(self) -> f64 {
-        self.mant as f64 / 10f64.powi(self.scale as i32)
-    }
-
-    /// The plain (never exponent) decimal lexical at this value's scale: scale 0 prints
-    /// as an integer ("3"); otherwise exactly `scale` fraction digits ("3.0", "0.05").
-    fn lexical(self) -> String {
-        let mag = self.mant.unsigned_abs().to_string();
-        let s = self.scale as usize;
-        let mut out = String::with_capacity(mag.len() + s + 2);
-        if self.mant < 0 {
-            out.push('-');
-        }
-        if s == 0 {
-            out.push_str(&mag);
-            return out;
-        }
-        if mag.len() > s {
-            out.push_str(&mag[..mag.len() - s]);
-        } else {
-            out.push('0');
-        }
-        out.push('.');
-        for _ in mag.len()..s {
-            out.push('0');
-        }
-        if mag.len() > s {
-            out.push_str(&mag[mag.len() - s..]);
-        } else {
-            out.push_str(&mag);
-        }
-        out
     }
 }
